fftTrustTest.m:  Now calls formatScopeData for the directory that is passed as the first argument.  Returns freq vector, power vector, time vector, and voltage amplitude vector. formatScopeData.pl: Now takes in a directory, looks for .CSV files from scopes (it doesn't distinguish between non-scope .CSV files) and outputs three files - Numerical.csv, Settings.csv, Settings.conf(I think).  Numerical has just the numerical data from the scope.  Settings.csv has, in one column, the settings data with non-numerical settings set to -1.  Settings.conf has the settings in the form '<setting>=<value>'.  This makes reading in the numerical data and numerical settings easy with Matlab's csvread(); fourierAll.m: Looks through the current directory for any other directories and calls fftTrustTest() for all found directories.  It compiles the frequency analysis data from each directory, and then plots power (magnitude of complex fourier trans) vs frequency on two plots (one for lateral tests, one for vertical tests).  Because the scope data comes in at 40kHz, it only plots up to the last frequency for which a certain power is achieved. TODO:  Make separate plots for tests with mass on the trust, and mass not on the truss.  Also, make the power cutoff either an argument (as in, make this a function) or figure out a good metric for deciding a cutoff power.
analyze.m: Now has option parsing, with the only currently implemented option '-c' that stops the function from clearing all plots so that subsequent calls with '-c' will plot data on top of previous calls' plots (useful for transient response tests).  Also checks to see if there are transient/static test data files and only attempts to plot the type that is there.  Linear fits are also now on the |F| vs rpm, |M| vs rpm, |F| vs duty cycle, |M| vs duty cycle (the duty cycle plots are new).

workingAnalyzeCalls.m: If the directory structure is untouched, running this file will plot the four trials we have for our motor and plot the results on top of each other.

Getting tired of typing for this commit.  The StateSpace model directory has the beginning of a state step function for a full 3D Storque model (based on Daniel Mellinger's work).  There are some small changes in the simulink folder, but none are significant and none change the fact that the simulink model is horribly broken.

-Ian