StateSpaceModel/ProgressSummary.tex: Added explanation of motor response delay and how experimental data was used to obtain a tentative figure.  No figure included here yet because my transient tests plots are nasty.
Also added a section with plots of the thrust and moment data vs RPM each with a polynomial fit (ignoring the erroneous data).  These fit values are used in our model.

Thrust Analysis: No significant changes except for a plea for refactoring in the header.  This thing has become such an ugly repetivite beast that it really doesn't make sense for anyone but me to use it/refactor it.  If you attempt to, please branch so as not to break it's fragile state and prevent me from going back to get plots from it when needed.  Thanks, okay, ya.